requirements.txt
asgiref==3.8.1
c104==2.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psutil==6.1.1
PyJWT==2.10.1
pymodbus==3.8.2
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
